Confirm delete button implemented

+++ app/src/main/java/com/ponichTech/pswdManager/ui/items/all_password_items/AllPasswordItemsFragment.kt
@@ -5,6 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
@@ -26,15 +27,16 @@ import com.ponichTech.pswdManager.utils.Resource
 import com.ponichTech.pswdManager.utils.autoCleared
 import kotlinx.coroutines.launch
 
-class AllPasswordItemsFragment : Fragment(){
+class AllPasswordItemsFragment : Fragment() {
 
-    private var binding :FragmentAllPasswordsItemsBinding by autoCleared()
+    private var binding: FragmentAllPasswordsItemsBinding by autoCleared()
     private val sharedViewModel: SharedViewModel by activityViewModels()
 
     private val viewModel: FirebasePasswordItemsViewModel by viewModels {
         FirebasePasswordItemsViewModel.FirebasePasswordItemsViewModelFactory(
             AuthRepositoryFirebase(),
-            PasswordItemRepositoryFirebase())
+            PasswordItemRepositoryFirebase()
+        )
     }
 
     override fun onCreateView(
@@ -42,8 +44,8 @@ class AllPasswordItemsFragment : Fragment(){
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentAllPasswordsItemsBinding.inflate(inflater,container,false)
-        binding.fab.setOnClickListener{
+        binding = FragmentAllPasswordsItemsBinding.inflate(inflater, container, false)
+        binding.fab.setOnClickListener {
             findNavController().navigate(R.id.action_allItemsFragment_to_addItemFragment)
         }
         return binding.root
@@ -61,7 +63,8 @@ class AllPasswordItemsFragment : Fragment(){
                     sharedViewModel.setUserId(it)
                 }
             } else if (currentUserResource is Resource.Error) {
-                Toast.makeText(requireContext(), currentUserResource.message, Toast.LENGTH_SHORT).show()
+                Toast.makeText(requireContext(), currentUserResource.message, Toast.LENGTH_SHORT)
+                    .show()
             }
         }
 
@@ -76,18 +79,21 @@ class AllPasswordItemsFragment : Fragment(){
 
         }
         binding.recyclerView.layoutManager = LinearLayoutManager(requireContext())
-        binding.recyclerView.adapter = PasswordItemAdapter(object : PasswordItemAdapter.PassItemListener {
-            override fun onItemClicked(passItems: PasswordItem) {
-                Toast.makeText(requireContext(),passItems.serviceName,Toast.LENGTH_SHORT).show()
-            }
-            //LongClicked
-            override fun onItemLongClicked(passItems: PasswordItem) {
-                sharedViewModel.selectPasswordItem(passItems)
-                findNavController()
-                    .navigate(R.id.action_allItemsFragment_to_detailItemFragment)
-            }
-        })
+        binding.recyclerView.adapter =
+            PasswordItemAdapter(object : PasswordItemAdapter.PassItemListener {
+                override fun onItemClicked(passItems: PasswordItem) {
+                    Toast.makeText(requireContext(), passItems.serviceName, Toast.LENGTH_SHORT)
+                        .show()
+                }
 
+                //LongClicked
+                override fun onItemLongClicked(passItems: PasswordItem) {
+                    sharedViewModel.selectPasswordItem(passItems)
+                    findNavController()
+                        .navigate(R.id.action_allItemsFragment_to_detailItemFragment)
+                }
+            })
+        ////////////////////////////////////////////////////////////////////////////////////////
         sharedViewModel.userId.observe(viewLifecycleOwner, Observer { userId ->
             if (userId != null) {
                 // Fetch or update data based on the logged-in userId
@@ -97,113 +103,84 @@ class AllPasswordItemsFragment : Fragment(){
             }
         })
 
-        ////////////////////////////////////////////////////////////////////////////////////////
+
 
         viewModel.passwordItems.observe(viewLifecycleOwner) {
-            when(it) {
+            when (it) {
                 is Resource.Loading -> {
                     binding.progressBar.isVisible = true
                     binding.fab.isEnabled = false
                 }
+
                 is Resource.Success -> {
                     binding.progressBar.isVisible = false
                     binding.fab.isEnabled = true
                     (binding.recyclerView.adapter as PasswordItemAdapter).setPasswordItems(it.data!!)
                 }
+
                 is Resource.Error -> {
                     binding.progressBar.isVisible = false
                     binding.fab.isEnabled = true
-                    Toast.makeText(requireContext(),it.message,Toast.LENGTH_SHORT).show()
+                    Toast.makeText(requireContext(), it.message, Toast.LENGTH_SHORT).show()
                 }
 
             }
         }
 
         viewModel.addPassStatus.observe(viewLifecycleOwner) {
-            when(it) {
+            when (it) {
                 is Resource.Loading -> {
                     binding.progressBar.isVisible = true
                 }
+
                 is Resource.Success -> {
                     binding.progressBar.isVisible = false
-                    Snackbar.make(binding.coordinator,"Item Added!",Snackbar.LENGTH_SHORT).show()
+                    Snackbar.make(binding.coordinator, "Item Added!", Snackbar.LENGTH_SHORT).show()
                 }
+
                 is Resource.Error -> {
                     binding.progressBar.isVisible = false
-                    Toast.makeText(requireContext(),it.message,Toast.LENGTH_SHORT).show()
+                    Toast.makeText(requireContext(), it.message, Toast.LENGTH_SHORT).show()
                 }
 
             }
         }
-
 
         viewModel.deletePassStatus.observe(viewLifecycleOwner) {
-            when(it) {
+            when (it) {
                 is Resource.Loading -> {
                     binding.progressBar.isVisible = true
                 }
+
                 is Resource.Success -> {
                     binding.progressBar.isVisible = false
-                    Snackbar.make(binding.coordinator,"Item Deleted!",Snackbar.LENGTH_SHORT)
+                    Snackbar.make(binding.coordinator, "Item Deleted!", Snackbar.LENGTH_SHORT)
                         .setAction("Undo") {
-                            Toast.makeText(requireContext(),"For you to implement",Toast.LENGTH_SHORT).show()
+                            Toast.makeText(
+                                requireContext(),
+                                "For you to implement",
+                                Toast.LENGTH_SHORT
+                            ).show()
                         }.show()
                 }
+
                 is Resource.Error -> {
                     binding.progressBar.isVisible = false
-                    Toast.makeText(requireContext(),it.message,Toast.LENGTH_SHORT).show()
+                    Toast.makeText(requireContext(), it.message, Toast.LENGTH_SHORT).show()
                 }
 
             }
         }
-/*////////////////////////////////////////////////////////////////////////////////////////
 
-//        val menuHost: MenuHost = requireActivity()
-//        menuHost.addMenuProvider(object : MenuProvider {
-//            override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
-//                menuInflater.inflate(R.menu.main_menu, menu)
-//            }
-//            override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
-//                return when (menuItem.itemId) {
-//                    R.id.action_delete -> {
-//                        val builder = AlertDialog.Builder(requireContext())
-//                        builder.setTitle(getString(R.string.confirm_delete))
-//                            .setMessage(getString(R.string.confirmation_delete_all))
-//                            .setPositiveButton(getString(R.string.delete_exclamation_mark)) { _, _ ->
-////                                viewModel.deleteAllItems()
-//                                Toast.makeText(
-//                                    requireContext(),
-//                                    getString(R.string.all_items_have_been_deleted),
-//                                    Toast.LENGTH_SHORT
-//                                ).show()
-//                            }
-//                            .setNegativeButton(android.R.string.cancel, null)
-//                            .show()
-//                        true
-//                    }
-//                    R.id.log_out -> {
-//                        val builder = AlertDialog.Builder(requireContext())
-//                        builder.setTitle(getString(R.string.confirm_logout))
-//                            .setMessage(getString(R.string.confirm_logout))
-//                            .setPositiveButton(getString(R.string.logout)) { _, _ ->
-//                                viewModel.signOut()
-//                                findNavController().navigate(R.id.action_allItemsFragment_to_loginFragment2)
-//                                Toast.makeText(requireContext(),getString(R.string.logout),Toast.LENGTH_SHORT).show()
-//                            }
-//                            .setNegativeButton(android.R.string.cancel, null)
-//                            .show()
-//                        true
-//                    }
-//                    else -> false
-//                }
-//            }
-//        }, viewLifecycleOwner, Lifecycle.State.RESUMED)*/
-
-        ItemTouchHelper(object:ItemTouchHelper.Callback(){
+        ItemTouchHelper(object : ItemTouchHelper.Callback() {
             override fun getMovementFlags(
                 recyclerView: RecyclerView,
                 viewHolder: RecyclerView.ViewHolder
-            )= makeFlag(ItemTouchHelper.ACTION_STATE_SWIPE,ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT)
+            ) = makeFlag(
+                ItemTouchHelper.ACTION_STATE_SWIPE,
+                ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT
+            )
+
             override fun onMove(
                 recyclerView: RecyclerView,
                 viewHolder: RecyclerView.ViewHolder,
@@ -211,13 +188,21 @@ class AllPasswordItemsFragment : Fragment(){
             ): Boolean {
                 TODO("Not yet implemented")
             }
+
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                val item = (binding.recyclerView.adapter as PasswordItemAdapter).itemAt(viewHolder.adapterPosition)
-                viewModel.deletePassItem(item.id)
+                val item =
+                    (binding.recyclerView.adapter as PasswordItemAdapter).itemAt(viewHolder.adapterPosition)
+                val builder = AlertDialog.Builder(requireContext())
+                builder.setTitle(getString(R.string.confirm_delete))
+                    .setMessage(getString(R.string.confirmation_delete_all))
+                    .setPositiveButton(getString(R.string.delete_exclamation_mark)) { _, _ ->
+                        viewModel.deletePassItem(item.id)
+                    }
+                    .setNegativeButton(android.R.string.cancel, null)
+                    .show()
             }
         }).attachToRecyclerView(binding.recyclerView)
     }
-
 
 
 }
